refactor: controller, service command로 수정
