+++ src/backend/state-service/src/main/java/capstone/stateservice/domain/userstate/service/command/UserStateCommandService.java
@@ -1,9 +1,9 @@
-package capstone.stateservice.domain.userstate.service;
+package capstone.stateservice.domain.userstate.service.command;
 
 import capstone.stateservice.global.common.dto.DataResponseDto;
 import capstone.stateservice.infra.kafka.consumer.state.dto.ConnectionStateInfo;
 
-public interface UserStateService {
+public interface UserStateCommandService {
 
     DataResponseDto<Long> saveUserConnectionState(ConnectionStateInfo connectionStateInfo);
 }
