+++ src/backend/state-service/src/main/java/capstone/stateservice/infra/kafka/consumer/state/connectionStateInfoConsumer.java
@@ -1,6 +1,6 @@
 package capstone.stateservice.infra.kafka.consumer.state;
 
-import capstone.stateservice.domain.userstate.service.UserStateService;
+import capstone.stateservice.domain.userstate.service.command.UserStateCommandService;
 import capstone.stateservice.infra.kafka.consumer.state.dto.ConnectionStateInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,10 +12,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class connectionStateInfoConsumer {
 
-    private final UserStateService userStateService;
+    private final UserStateCommandService userStateCommandService;
 
     @KafkaListener(topics = "${spring.kafka.topic.connection-state-info}", groupId = "${spring.kafka.consumer.group-id.connection-state-info}", containerFactory = "connectionStateInfoListenerContainerFactory")
     public void connectionStateInfoListener(ConnectionStateInfo connectionStateInfo) {
-        userStateService.saveUserConnectionState(connectionStateInfo);
+        userStateCommandService.saveUserConnectionState(connectionStateInfo);
     }
 }
